chore: rename Vectorizer -> VectorConfig and move static factories under the new namespace
