+++ src/main/java/io/weaviate/client6/v1/api/collections/config/UpdateCollectionRequest.java
@@ -10,7 +10,7 @@ import io.weaviate.client6.v1.api.collections.Generative;
 import io.weaviate.client6.v1.api.collections.InvertedIndex;
 import io.weaviate.client6.v1.api.collections.Replication;
 import io.weaviate.client6.v1.api.collections.Reranker;
-import io.weaviate.client6.v1.api.collections.Vectorizer;
+import io.weaviate.client6.v1.api.collections.VectorConfig;
 import io.weaviate.client6.v1.internal.ObjectBuilder;
 import io.weaviate.client6.v1.internal.json.JSON;
 import io.weaviate.client6.v1.internal.rest.Endpoint;
@@ -95,13 +95,13 @@ public record UpdateCollectionRequest(CollectionConfig collection) {
       return this;
     }
 
-    public final Builder vectors(Map<String, Vectorizer> vectors) {
+    public final Builder vectors(Map<String, VectorConfig> vectors) {
       this.newCollection.vectorConfig(vectors);
       return this;
     }
 
     @SafeVarargs
-    public final Builder vectors(Map.Entry<String, Vectorizer>... vectors) {
+    public final Builder vectors(Map.Entry<String, VectorConfig>... vectors) {
       this.newCollection.vectorConfig(vectors);
       return this;
     }
