+++ src/main/java/io/weaviate/client6/v1/api/collections/vectorizers/Multi2VecClipVectorizer.java
@@ -8,8 +8,8 @@ import java.util.function.Function;
 
 import com.google.gson.annotations.SerializedName;
 
+import io.weaviate.client6.v1.api.collections.VectorConfig;
 import io.weaviate.client6.v1.api.collections.VectorIndex;
-import io.weaviate.client6.v1.api.collections.Vectorizer;
 import io.weaviate.client6.v1.internal.ObjectBuilder;
 
 public record Multi2VecClipVectorizer(
@@ -22,7 +22,7 @@ public record Multi2VecClipVectorizer(
     /** Weights of the included properties. */
     @SerializedName("weights") Weights weights,
     /** Vector index configuration. */
-    VectorIndex vectorIndex) implements Vectorizer {
+    VectorIndex vectorIndex) implements VectorConfig {
 
   private static record Weights(
       /**
@@ -38,8 +38,8 @@ public record Multi2VecClipVectorizer(
   }
 
   @Override
-  public Vectorizer.Kind _kind() {
-    return Vectorizer.Kind.MULTI2VEC_CLIP;
+  public VectorConfig.Kind _kind() {
+    return VectorConfig.Kind.MULTI2VEC_CLIP;
   }
 
   @Override
